+++ src/Map/DoubleIndex.php
@@ -241,7 +241,6 @@ final class DoubleIndex implements Implementation
      */
     public function merge(Implementation $map): self
     {
-        /** @psalm-suppress MixedArgument For some reason it no longer recognize templates for $key and $value */
         return $map->reduce(
             $this,
             static fn(self $carry, $key, $value): self => ($carry)($key, $value),
@@ -276,11 +275,13 @@ final class DoubleIndex implements Implementation
     }
 
     /**
+     * @template I
      * @template R
-     * @param R $carry
-     * @param callable(R, T, S): R $reducer
      *
-     * @return R
+     * @param I $carry
+     * @param callable(I|R, T, S): R $reducer
+     *
+     * @return I|R
      */
     public function reduce($carry, callable $reducer)
     {