+++ src/Map/Implementation.php
@@ -149,11 +149,13 @@ interface Implementation extends \Countable
     /**
      * Reduce the map to a single value
      *
+     * @template I
      * @template R
-     * @param R $carry
-     * @param callable(R, T, S): R $reducer
      *
-     * @return R
+     * @param I $carry
+     * @param callable(I|R, T, S): R $reducer
+     *
+     * @return I|R
      */
     public function reduce($carry, callable $reducer);