+++ tests/Sequence/DeferTest.php
@@ -653,16 +653,14 @@ class DeferTest extends TestCase
     {
         $loaded = false;
         $a = new Defer((static function() use (&$loaded) {
-            yield 1;
-            yield 2;
-            yield 3;
-            yield 4;
+            yield from [1, 2];
+            yield from [3, 4];
             $loaded = true;
         })());
         $b = $a->reverse();
 
         $this->assertFalse($loaded);
-        $this->assertSame([1, 2, 3, 4], \iterator_to_array($a->iterator()));
+        $this->assertSame([1, 2, 3, 4], $a->toSequence()->toList());
         $this->assertInstanceOf(Defer::class, $b);
         $this->assertSame([4, 3, 2, 1], \iterator_to_array($b->iterator()));
         $this->assertTrue($loaded);